+++ mongoengine_plus/models/__init__.py
@@ -1,3 +1,4 @@
-__all__ = ['BaseModel']
+__all__ = ['BaseModel', 'uuid_field']
 
 from .base import BaseModel
+from .helpers import uuid_field
