+++ mongoengine_plus/models/base.py
@@ -1,6 +1,6 @@
 from typing import ClassVar, Dict
 
-from mongoengine_plus.types.helpers import mongo_to_dict
+from .helpers import mongo_to_dict
 
 
 class BaseModel:
