+++ mongoengine_plus/types/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['EnumField']
+
+from .enum_field import EnumField
